revert(provider-graphql): codegen-unsafe 수동 분해 원복
codegen.ts 의 target 이 src/index.ts (plugins: ['typescript']) 로 설정되어 있어 Wave 3 에서 수동으로 쪼갠 core/magazine/product/operations.ts 는 pnpm codegen 실행 시 고아 stale 파일이 되고, src/index.ts 는 다시 단일 생성본으로 덮어써짐. 스키마 변경 시에도 수동 split 이 따라가지 못해 type drift 위험.

외부 consumer 43 건은 전부 @darun/provider-graphql 루트 path 만 사용하므로 domain 별 split 은 consumer 관점에서 invisible — 내부 조직화 외 실익 없음.

- src/core.ts, src/magazine.ts, src/product.ts, src/operations.ts 삭제
- src/index.ts 를 trunk 원본 (codegen typescript plugin 생성본) 으로 복원
- eslint.config.mjs 에 ignores: ['libs/shared/provider-graphql/src/index.ts'] rule 복원 (자동생성 파일 lint 제외)
